requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.5
django-appconf==1.1.0
django-compressor==4.5.1
django-debug-toolbar==5.0.1
django-htmlmin==0.11.0
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
s3transfer==0.10.4
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
yarl==1.18.3
